Move gensx to dependency of create-gensx.
