+++ packages/gensx-cli/src/templates/projects/typescript/README.md.template
