+++ packages/gensx-cli/src/templates/projects/typescript/nodemon.json.template
