+++ packages/gensx-cli/src/templates/projects/typescript/package.json.template
