+++ packages/gensx-cli/src/templates/projects/typescript/src/index.tsx.template
