+++ packages/gensx-cli/src/templates/projects/typescript/template.json
