+++ packages/gensx-cli/src/templates/projects/typescript/tsconfig.json.template
