+++ turbo.json
@@ -11,11 +11,6 @@
       "dependsOn": ["^build"],
       "outputs": []
     },
-    "dev": {
-      "dependsOn": ["^build"],
-      "cache": false,
-      "persistent": true
-    },
     "clean": {
       "cache": false
     },
